requirements.txt
Pillow
pixels2svg